Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.2
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
motor==3.3.2
pymongo==4.6.0
beautifulsoup4==4.12.2
httpx==0.25.2
aiohttp==3.9.1
textstat==0.7.3
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
python-multipart==0.0.6
lxml==4.9.3
fake-useragent==1.4.0
python-slugify==8.0.1
